+++ src/Java/myservicebus-tasks/src/main/java/com/myservicebus/tasks/CancellationRegistration.java
@@ -1,0 +1,7 @@
+package com.myservicebus.tasks;
+
+@FunctionalInterface
+public interface CancellationRegistration extends AutoCloseable {
+    @Override
+    void close();
+}